♻️ Validator class 형으로 변경
